修正了若干错误，形成了一个尚且可用的版本（为了争取在笔记本没电之前把原型搞出来，写了大量的硬编码），命令提示符真的是好东西，可以告诉你一次处理已经完毕了。一开始没写，傻傻的等了半天。

+++ DataFetcher.cs
@@ -62,8 +62,8 @@ namespace BillyNaCl.QQGroupToolkit
                 string role = memberJsonNode?["role"]?.ToString()!;
                 long? ShutUpEndTime = memberJsonNode?["shut_up_end_time"]?.GetValue<long>();
                 string title = memberJsonNode?["title"]?.ToString()!;
-                int UserId = (int)(memberJsonNode?["user_id"]?.GetValue<int>())!;
-                int GroupId = (int)(memberJsonNode?["group_id"]?.GetValue<int>())!;
+                long UserId = (long)(memberJsonNode?["user_id"]?.GetValue<long>())!;
+                long GroupId = (int)(memberJsonNode?["group_id"]?.GetValue<long>())!;
                 GroupMember m = new(Gender, JoinTime, LastSpeakTime, LevelInGroup, NameInGroup, NickName, role, ShutUpEndTime, title, UserId, GroupId);
                 result.Add(m);
             }
